plate-recognition/requirements.txt
ultralytics==8.0.114
pandas==2.0.2
opencv-python==4.7.0.72
numpy==1.24.3
scipy==1.10.1
easyocr==1.7.0
paho-mqtt==1.6.1
urllib3==2.0.5